Move driver to own subdirectory
